+++ src/@types/DTAParser.ts
@@ -1,8 +1,17 @@
-export type SortByOptionsTypes = 'name' | 'artist' | 'artist_set' | 'song_id'
+import { UpdateDataOptions } from '../core'
+
+export type SortByOptionsTypes =
+    | 'id'
+    | 'name'
+    | 'artist'
+    | 'artist_set'
+    | 'song_id'
 
 export interface DTAParserOptions {
     /** Changes the sorting of the songs. Valid values are:
      *
+     * - - - -
+     * * `id` => Sorts by the songs' unique string ID.
      * - - - -
      * * `name` => Sorts by the songs' title (excluding the leading article, if any).
      * - - - -
@@ -15,9 +24,20 @@ export interface DTAParserOptions {
      */
     sortBy?: SortByOptionsTypes
     /**
+     * Set this to `true` only if you're parsing official Harmonix songs DTAs.
+     *
      * If you're parsing a DTA file from official pre-RB3 songs, the parser will
      * try to seek additional information from the updates.
+     *
+     * Also, It'll put Harmonix as author and the multitracks value to `true` on all songs.
      * @default false
      */
-    applyUpdates?: boolean
+    harmonixSongs?: boolean
+    /**
+     * Applies direct values updates for any song parsed based on the song's unique string ID.
+     */
+    update?: {
+        [id: string]: UpdateDataOptions
+    }
+    updateAll?: Pick<UpdateDataOptions, 'author'>
 }